requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.0.0
pypdf>=4.0.0
pdfplumber>=0.11.0
flask>=3.0.0
flask-sqlalchemy>=3.1.1
Flask-Limiter>=3.5.0
Flask-Migrate>=4.1.0
werkzeug>=3.0.0

# Authentication and Security
bcrypt>=4.1.0
PyJWT>=2.8.0
Flask-Limiter>=3.5.0

# Database optimization and management
alembic>=1.13.0  # Database migrations
sqlalchemy-utils>=0.41.0  # Advanced database utilities

# Caching (elegir una opción según preferencia)
# Opción 1: Redis (más potente, requiere servidor Redis)
redis>=5.0.0

# Opción 2: DiskCache (alternativa sin servidor externo)
# diskcache>=5.6.0

# Performance monitoring (opcional)
flask-caching>=2.1.0  # Cache Flask routes
